+++ lib/flutter_awesome_reels.dart
@@ -1,8 +1,7 @@
-library flutter_awesome_reels;
-
 // Core exports
 export 'src/awesome_reels.dart';
-export 'src/models/reel_model.dart' show ReelModel, ReelUser, ReelAudio, VideoSource, VideoFormat;
+export 'src/models/reel_model.dart'
+    show ReelModel, ReelUser, ReelAudio, VideoSource, VideoFormat;
 export 'src/models/reel_config.dart';
 export 'src/models/reel_analytics.dart';
 export 'src/widgets/reel_video_player.dart';
@@ -14,4 +13,3 @@ export 'src/services/analytics_service.dart';
 export 'src/services/streaming_service.dart';
 export 'src/controllers/reel_controller.dart';
 export 'src/utils/reel_utils.dart';
-